services/clustering-service/requirements.txt
# services/clustering-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aio_pika>=9.3.0
numpy>=1.26.0
httpx>=0.25.0
aiohttp>=3.9.0
prometheus-client>=0.19.0
python-jose[cryptography]>=3.3.0

# OpenAI Embeddings for semantic topic search
openai>=1.0.0
cachetools>=5.3.0

# Batch Clustering (UMAP+HDBSCAN) - Epic 2
celery>=5.3.0
redis>=5.0.0
umap-learn>=0.5.5
hdbscan>=0.8.33
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# pgvector for vector similarity search
pgvector>=0.2.0

# Alembic for migrations
alembic>=1.13.0

# Shared library
news-intelligence-common
